requirements.txt
streamlit==1.25.0
request
